Fix delete, implement get_option_columns and fix grouping
Get the actual namespace of the reftrack by quering the namespace
attribute and then combining it with the namespace of the reftrack node
itself. Then delete that namespace. Do not use force flag on the
removeReference. It will somehow also remove the parent namespace.

Show the third (the version) column for the second option level for
assets. Before we showed column 0 of the taskfileinfo and that was the
name of the element.

Only group top level transforms (assemblies). Not all dag nodes.
Grouping all dagnodes tries to flatten the hierarchy and cannot be done
for read only nodes anyway (references).