Fix import then reference bug
If you import a file, then reference it, the filenames of the references
are buggy and you cannot use them, e.g. for quering the reference node.
Also you cannot use them to delete a reference, because they will not have
copy numbers if you reference the file multiple times.